(fix) изменил WeatherPresenter
