+++ app/src/main/java/com/example/user/weatherapp/presenter/IWeatherPresenter.java
@@ -1,0 +1,3 @@
+public interface IWeatherView extends MvpView {
+    void loadWeather();
+}
